+++ modules/src/main/java/org/hortonmachine/modules/GeopaparazziSpatialiteCreator.java
@@ -293,27 +293,27 @@ public class GeopaparazziSpatialiteCreator extends HMModel {
             PointSymbolizerWrapper psw = (PointSymbolizerWrapper) geometrySymbolizersWrapper;
 
             gpStyle.shape = psw.getMarkName();
-            gpStyle.size = pSizeFactor * getFloat(psw.getSize(), gpStyle.size);
+            gpStyle.size = pSizeFactor * getDouble(psw.getSize(), gpStyle.size);
 
-            gpStyle.width = pSizeFactor * getFloat(psw.getStrokeWidth(), gpStyle.width);
-            gpStyle.strokealpha = getFloat(psw.getStrokeOpacity(), gpStyle.strokealpha);
+            gpStyle.width = pSizeFactor * getDouble(psw.getStrokeWidth(), gpStyle.width);
+            gpStyle.strokealpha = getDouble(psw.getStrokeOpacity(), gpStyle.strokealpha);
             gpStyle.strokecolor = getString(psw.getStrokeColor(), null);
 
-            gpStyle.fillalpha = getFloat(psw.getFillOpacity(), gpStyle.fillalpha);
+            gpStyle.fillalpha = getDouble(psw.getFillOpacity(), gpStyle.fillalpha);
             gpStyle.fillcolor = getString(psw.getFillColor(), null);
         } else if (geometrySymbolizersWrapper instanceof PolygonSymbolizerWrapper) {
             PolygonSymbolizerWrapper psw = (PolygonSymbolizerWrapper) geometrySymbolizersWrapper;
 
-            gpStyle.width = pSizeFactor * getFloat(psw.getStrokeWidth(), gpStyle.width);
-            gpStyle.strokealpha = getFloat(psw.getStrokeOpacity(), gpStyle.strokealpha);
+            gpStyle.width = pSizeFactor * getDouble(psw.getStrokeWidth(), gpStyle.width);
+            gpStyle.strokealpha = getDouble(psw.getStrokeOpacity(), gpStyle.strokealpha);
             gpStyle.strokecolor = getString(psw.getStrokeColor(), null);
-            gpStyle.fillalpha = getFloat(psw.getFillOpacity(), gpStyle.fillalpha);
+            gpStyle.fillalpha = getDouble(psw.getFillOpacity(), gpStyle.fillalpha);
             gpStyle.fillcolor = getString(psw.getFillColor(), null);
         } else if (geometrySymbolizersWrapper instanceof LineSymbolizerWrapper) {
             LineSymbolizerWrapper lsw = (LineSymbolizerWrapper) geometrySymbolizersWrapper;
 
-            gpStyle.width = pLinesWidthFactor * getFloat(lsw.getStrokeWidth(), gpStyle.width);
-            gpStyle.strokealpha = getFloat(lsw.getStrokeOpacity(), gpStyle.strokealpha);
+            gpStyle.width = pLinesWidthFactor * getDouble(lsw.getStrokeWidth(), gpStyle.width);
+            gpStyle.strokealpha = getDouble(lsw.getStrokeOpacity(), gpStyle.strokealpha);
             gpStyle.strokecolor = getString(lsw.getStrokeColor(), null);
         }
     }
@@ -347,10 +347,10 @@ public class GeopaparazziSpatialiteCreator extends HMModel {
         return gpStyle;
     }
 
-    private float getFloat( String value, float defaultValue ) {
-        float num = defaultValue;
+    private double getDouble( String value, double defaultValue ) {
+        double num = defaultValue;
         try {
-            num = Float.parseFloat(value);
+            num = Double.parseDouble(value);
         } catch (Exception e) {
             // ignore and get default
         }
